+++ src/frequenz/channels/_anycast.py
@@ -10,6 +10,8 @@ from asyncio import Condition
 from collections import deque
 from typing import Generic, TypeVar
 
+from typing_extensions import override
+
 from ._exceptions import ChannelClosedError
 from ._generic import ChannelMessageT
 from ._receiver import Receiver, ReceiverStoppedError
@@ -320,6 +322,7 @@ class _Sender(Sender[_T]):
         self._channel: Anycast[_T] = channel
         """The channel that this sender belongs to."""
 
+    @override
     async def send(self, message: _T, /) -> None:
         """Send a message across the channel.
 
@@ -388,8 +391,12 @@ class _Receiver(Receiver[_T]):
         self._channel: Anycast[_T] = channel
         """The channel that this receiver belongs to."""
 
+        self._closed: bool = False
+        """Whether the receiver is closed."""
+
         self._next: _T | type[_Empty] = _Empty
 
+    @override
     async def ready(self) -> bool:
         """Wait until the receiver is ready with a message or an error.
 
@@ -405,6 +412,9 @@ class _Receiver(Receiver[_T]):
         if self._next is not _Empty:
             return True
 
+        if self._closed:
+            return False
+
         # pylint: disable=protected-access
         while len(self._channel._deque) == 0:
             if self._channel._closed:
@@ -417,6 +427,7 @@ class _Receiver(Receiver[_T]):
         # pylint: enable=protected-access
         return True
 
+    @override
     def consume(self) -> _T:
         """Return the latest message once `ready()` is complete.
 
@@ -431,6 +442,9 @@ class _Receiver(Receiver[_T]):
         ):
             raise ReceiverStoppedError(self) from ChannelClosedError(self._channel)
 
+        if self._next is _Empty and self._closed:
+            raise ReceiverStoppedError(self)
+
         assert (
             self._next is not _Empty
         ), "`consume()` must be preceded by a call to `ready()`"
@@ -440,6 +454,14 @@ class _Receiver(Receiver[_T]):
         self._next = _Empty
 
         return next_val
+
+    @override
+    def close(self) -> None:
+        """Close this receiver.
+
+        After closing, the receiver will not be able to receive any more messages.
+        """
+        self._closed = True
 
     def __str__(self) -> str:
         """Return a string representation of this receiver."""